+++ source/system/Entry.js
