+++ source/system/Flattener.js
