+++ source/system/InigoGenerator.js
