+++ source/system/SharedWorkspace.js
