+++ source/system/commands/BaseCommand.js
