+++ source/system/commands/CommentCommand.js
