+++ source/system/commands/CreateEntryCommand.js
