+++ source/system/commands/CreateGroupCommand.js
