+++ source/system/commands/DeleteArchiveAttributeCommand.js
