+++ source/system/commands/DeleteEntryCommand.js
