+++ source/system/commands/DeleteEntryMetaCommand.js
