+++ source/system/commands/DeleteGroupAttributeCommand.js
