+++ source/system/commands/DeleteGroupCommand.js
