+++ source/system/commands/FormatCommand.js
