+++ source/system/commands/MoveGroupCommand.js
