+++ source/system/commands/SetArchiveAttributeCommand.js
