+++ source/system/commands/SetEntryAttributeCommand.js
