+++ source/system/commands/SetEntryMetaCommand.js
