+++ source/system/commands/SetEntryPropertyCommand.js
