+++ source/system/commands/SetGroupAttributeCommand.js
